+++ src/Form/FichierProjetType.php
@@ -2,15 +2,10 @@
 
 namespace App\Form;
 
-use App\Entity\Fichier;
 use App\Entity\FichierProjet;
-use Symfony\Bridge\Doctrine\Form\Type\EntityType;
 use Symfony\Component\Form\FormBuilderInterface;
 use Symfony\Component\OptionsResolver\OptionsResolver;
 use Symfony\Component\Form\AbstractType;
-use Symfony\Component\Form\Extension\Core\Type\TextareaType;
-use Symfony\Component\Form\Extension\Core\Type\FileType;
-use Symfony\Component\Form\Extension\Core\Type\SubmitType;
 
 class FichierProjetType extends AbstractType
 {
@@ -19,16 +14,6 @@ class FichierProjetType extends AbstractType
         $builder
             ->add('fichier', FichierType::class, [
                 'label' => false,
-            ])
-            ->add('description', TextareaType::class, [
-                'label' => 'Description',
-                'required' => false,
-            ])
-            ->add('submit', SubmitType::class, [
-                'label' => 'Soumettre',
-                'attr' => [
-                    'class' => 'mt-5 btn btn-success'
-                ]
             ])
         ;
     }