+++ src/Form/ProjetFichierProjetsType.php
@@ -2,17 +2,17 @@
 
 namespace App\Form;
 
-use App\Entity\FaitMarquant;
+use App\Entity\Projet;
 use App\Form\Custom\FichierProjetsType;
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\Event\SubmitEvent;
+use Symfony\Component\Form\Extension\Core\Type\SubmitType;
 use Symfony\Component\Form\FormBuilderInterface;
-use Symfony\Component\OptionsResolver\OptionsResolver;
-use Symfony\Component\Form\Extension\Core\Type\TextareaType;
 use Symfony\Component\Form\FormEvents;
+use Symfony\Component\OptionsResolver\OptionsResolver;
 use Symfony\Component\Security\Core\Authentication\Token\Storage\TokenStorageInterface;
 
-class FaitMarquantType extends AbstractType
+class ProjetFichierProjetsType extends AbstractType
 {
     private TokenStorageInterface $tokenStorage;
 
@@ -24,18 +24,9 @@ class FaitMarquantType extends AbstractType
     public function buildForm(FormBuilderInterface $builder, array $options)
     {
         $builder
-            ->add('titre', null, [
-                'attr' => ['class' => 'form-control-lg'],
-            ])
-            ->add('description', TextareaType:: class, [
-                'attr' => [
-                    'class' => 'text-justify',
-                    'rows' => 9
-                ]
-            ])
             ->add('fichierProjets', FichierProjetsType::class, [
-                'projet' => $builder->getData()->getProjet(),
-                'label' => 'Fichiers joints',
+                'projet' => $builder->getData(),
+                'label' => false,
             ])
             ->addEventListener(FormEvents::SUBMIT, [$this, 'setFichierProjetFaitMarquant'])
         ;
@@ -43,16 +34,14 @@ class FaitMarquantType extends AbstractType
 
     public function setFichierProjetFaitMarquant(SubmitEvent $event)
     {
-        $faitMarquant = $event->getData();
+        $projet = $event->getData();
 
-        foreach ($faitMarquant->getFichierProjets() as $fichierProjet) {
+        foreach ($projet->getFichierProjets() as $fichierProjet) {
             if (null !== $fichierProjet->getId()) {
                 continue;
             }
 
             $fichierProjet
-                ->setProjet($faitMarquant->getProjet())
-                ->setFaitMarquant($faitMarquant)
                 ->setUploadedBy($this->tokenStorage->getToken()->getUser())
             ;
         }
@@ -61,7 +50,7 @@ class FaitMarquantType extends AbstractType
     public function configureOptions(OptionsResolver $resolver)
     {
         $resolver->setDefaults([
-            'data_class' => FaitMarquant::class,
+            'data_class' => Projet::class,
         ]);
     }
 }