+++ CHANGELOG.md
@@ -5,6 +5,8 @@ Tous les changements apportés au projet sont suivis dans ce fichier.
 Le format est basé sur [Tenez un Changelog](https://keepachangelog.com/fr/1.0.0/).
 
 ## [À venir dans la prochaine version]
+### Ajouté
+- Permet de joindre un fichier à un fait marquant
 ### Modifié
 - Refactorise l'entité FichierProjet dans Fichier pour être plus flexible sur les fichiers
 - Utilise flysystem comme couche d'abstraction du système de fichier